requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python==4.8.1.78
Pillow==10.0.1
numpy<2.0
python-dotenv==1.0.0